Bluetooth refactor to use SLC generation
