+++ examples/lighting-app/efr32/src/AppTask.cpp
@@ -40,8 +40,6 @@
 #include <setup_payload/QRCodeSetupPayloadGenerator.h>
 #include <setup_payload/SetupPayload.h>
 
-#include <platform/EFR32/freertos_bluetooth.h>
-
 #include <lib/support/CodeUtils.h>
 
 #include <platform/CHIPDeviceLayer.h>
