+++ src/platform/EFR32/BUILD.gn
@@ -57,8 +57,6 @@ static_library("EFR32") {
     "PlatformManagerImpl.cpp",
     "PlatformManagerImpl.h",
     "SystemPlatformConfig.h",
-    "freertos_bluetooth.c",
-    "freertos_bluetooth.h",
     "gatt_db.c",
     "gatt_db.h",
   ]
